+++ exporters/otlp/otlplog/otlploghttp/client.go
@@ -278,6 +278,7 @@ func (c *httpClient) newRequest(ctx context.Context, body []byte) (request, erro
 	case NoCompression:
 		r.ContentLength = int64(len(body))
 		req.bodyReader = bodyReader(body)
+		req.GetBody = bodyReaderErr(body)
 	case GzipCompression:
 		// Ensure the content length is not used.
 		r.ContentLength = -1
@@ -298,6 +299,7 @@ func (c *httpClient) newRequest(ctx context.Context, body []byte) (request, erro
 		}
 
 		req.bodyReader = bodyReader(b.Bytes())
+		req.GetBody = bodyReaderErr(body)
 	}
 
 	return req, nil
@@ -307,6 +309,13 @@ func (c *httpClient) newRequest(ctx context.Context, body []byte) (request, erro
 func bodyReader(buf []byte) func() io.ReadCloser {
 	return func() io.ReadCloser {
 		return io.NopCloser(bytes.NewReader(buf))
+	}
+}
+
+// bodyReaderErr returns a closure returning a new reader for buf.
+func bodyReaderErr(buf []byte) func() (io.ReadCloser, error) {
+	return func() (io.ReadCloser, error) {
+		return io.NopCloser(bytes.NewReader(buf)), nil
 	}
 }